Splits config and utils
