+++ sonyci/config.py
@@ -1,0 +1,2 @@
+BASE_URL = 'https://api.cimediacloud.com/'
+TOKEN_URL = 'https://api.cimediacloud.com/oauth2/token'
